implement some optimization techniques

+++ src/components/Forms/FormsPage.jsx
@@ -6,6 +6,7 @@ import Form from "./Form.jsx";
 import FormView from "./FormView.jsx";
 import FormsHeader from "./FormsHeader.jsx";
 import toast from "react-hot-toast";
+import useConfirm from "../../Hooks/useConfirm.jsx";
 
 function FormsPage() {
     const [forms, setForms] = useState([]);
@@ -15,6 +16,7 @@ function FormsPage() {
     const [showForm, setShowForm] = useState(false);
     const [formsToDelete, setFormsToDelete] = useState([]);
     const [showDelete, setShowDelete] = useState(false);
+    const [confirm, Confirmation] = useConfirm();
 
     useEffect(() => {
         (async () => {
@@ -63,6 +65,12 @@ function FormsPage() {
     }
 
     async function deleteForms() {
+        const result = await confirm("Are you sure you want to delete forms?");
+
+        if (!result) {
+            return;
+        }
+
         let id = toast.loading("Deleting Forms...");
         let token = await getToken();
         let response = await deleteForm(formsToDelete, token);
@@ -76,6 +84,13 @@ function FormsPage() {
     }
 
     async function deleteAll() {
+        console.log("delete all fun called")
+        const result = await confirm("Are you sure you want to delete All forms?");
+
+        if (!result) {
+            return;
+        }
+
         let id = toast.loading("Deleting Forms...");
         setFormsToDelete([]);
         let ids = forms.map(form => form.id);
@@ -95,6 +110,7 @@ function FormsPage() {
             {forms.map((form, idx) => <Form key={form.id} form={form} idx={idx} addForm={addForm}
                                             removeForm={removeForm}/>)}
             <FormView show={showForm} form={selectedForm} onClose={() => setShowForm(false)}/>
+            {Confirmation}
         </div>
     );
 }
